+++ app/pkg/usecases/hsmconnection/types.go
@@ -4,7 +4,6 @@ import (
 	"log/slog"
 
 	"github.com/lfdt-smoot/signare/app/pkg/entities"
-	"github.com/lfdt-smoot/signare/app/pkg/entities/address"
 	"github.com/lfdt-smoot/signare/app/pkg/usecases/hsmconnector"
 	"github.com/lfdt-smoot/signare/app/pkg/usecases/hsmslot"
 )
@@ -31,27 +30,6 @@ func (c HSMConnection) LogValue() slog.Value {
 	return slog.GroupValue(
 		slog.Any("slot", c.Slot),
 		slog.String("moduleKind", string(c.ModuleKind)),
+		slog.String("applicationDefaultChainId", c.ApplicationDefaultChainID.String()),
 	)
-}
-
-// SlotConfig defines the configuration for a particular slot.
-type SlotConfig struct {
-	AKV           []AKVConfig
-	LocalKeyVault *LocalKeyVaultConfig
-}
-
-// AKVConfig defines possible configurations for Azure Key Vault.
-type AKVConfig struct {
-	// KeyName is the name for the key to use in AKV
-	KeyName string
-	// KeyVersion is the version for the key to use in AKV
-	KeyVersion string
-	// KeyPublicAddress is the public address for the key to use in AKV
-	KeyPublicAddress string
-}
-
-// LocalKeyVaultConfig defines possible configurations for Local Key Vault.
-type LocalKeyVaultConfig struct {
-	// KeyStore holds the stored addresses and its private keys
-	KeyStore map[address.Address]string
 }